# Conflicts:
#	src/main/java/edu/ucdenver/ccp/knowtator/view/graph/GraphViewer.java
#	src/main/java/edu/ucdenver/ccp/knowtator/view/text/RectanglePainter.java
#	src/main/java/other/DnDTabbedPane.java
#	src/main/java/other/GraphPopupMenu.java
#	src/main/java/other/ListDialog.java
#	target/classes/META-INF/MANIFEST.MF
#	target/maven-status/maven-compiler-plugin/compile/default-compile/createdFiles.lst

+++ src/main/java/edu/ucdenver/ccp/knowtator/view/graph/GraphMenu.java
@@ -47,23 +47,48 @@ class GraphMenu extends JMenu {
         this.graphViewer = graphViewer;
 
         add(addNewGraphCommand());
+        add(renameGraphCommand());
         add(saveToImageCommand());
         add(deleteGraphCommand());
+    }
+
+    private JMenuItem renameGraphCommand() {
+        JMenuItem menuItem = new JMenuItem("Rename Graph");
+        menuItem.addActionListener(e -> {
+            if (graphViewer.getCurrentGraphComponent() != null) {
+                JTextField nameField = new JTextField(10);
+                nameField.setText(graphViewer.getCurrentGraphComponent().getName());
+                JPanel inputPanel = new JPanel();
+                inputPanel.add(new JLabel("New Name:"));
+                inputPanel.add(nameField);
+
+
+                int result = JOptionPane.showConfirmDialog(null, inputPanel,
+                        "Enter a new name for this graph space", JOptionPane.DEFAULT_OPTION);
+                if (result == JOptionPane.OK_OPTION) {
+                    graphViewer.renameCurrentGraph(nameField.getText());
+                }
+            }
+        });
+
+        return menuItem;
     }
 
     private JMenuItem saveToImageCommand() {
         JMenuItem menuItem = new JMenuItem("Save as PNG");
         menuItem.addActionListener(e -> {
-            JFileChooser fileChooser = new JFileChooser();
-            fileChooser.setCurrentDirectory(manager.getProjectManager().getProjectLocation());
-            FileFilter fileFilter = new FileNameExtensionFilter("PNG", "png");
-            fileChooser.setFileFilter(fileFilter);
-            if (fileChooser.showSaveDialog(null) == JFileChooser.APPROVE_OPTION) {
-                BufferedImage image = mxCellRenderer.createBufferedImage(graphViewer.getSelectedGraphComponent().getGraph(), null, 1, Color.WHITE, true, null);
-                try {
-                    ImageIO.write(image, "PNG", new File(fileChooser.getSelectedFile().getAbsolutePath()));
-                } catch (IOException e1) {
-                    e1.printStackTrace();
+            if (graphViewer.getCurrentGraphComponent() != null) {
+                JFileChooser fileChooser = new JFileChooser();
+                fileChooser.setCurrentDirectory(manager.getProjectManager().getProjectLocation());
+                FileFilter fileFilter = new FileNameExtensionFilter("PNG", "png");
+                fileChooser.setFileFilter(fileFilter);
+                if (fileChooser.showSaveDialog(null) == JFileChooser.APPROVE_OPTION) {
+                    BufferedImage image = mxCellRenderer.createBufferedImage(graphViewer.getCurrentGraphComponent().getGraph(), null, 1, Color.WHITE, true, null);
+                    try {
+                        ImageIO.write(image, "PNG", new File(fileChooser.getSelectedFile().getAbsolutePath()));
+                    } catch (IOException e1) {
+                        e1.printStackTrace();
+                    }
                 }
             }
         });
@@ -74,8 +99,10 @@ class GraphMenu extends JMenu {
     private JMenuItem deleteGraphCommand() {
         JMenuItem deleteGraphMenuItem = new JMenuItem("Delete graph");
         deleteGraphMenuItem.addActionListener(e -> {
-            if (JOptionPane.showConfirmDialog(null, "Are you sure you want to delete this graph?") == JOptionPane.YES_OPTION) {
-                graphViewer.deleteSelectedGraph();
+            if (graphViewer.getCurrentGraphComponent() != null) {
+                if (JOptionPane.showConfirmDialog(null, "Are you sure you want to delete this graph?") == JOptionPane.YES_OPTION) {
+                    graphViewer.deleteSelectedGraph();
+                }
             }
         });
